move tests to correct folder
